refactor(op): collapse duplicate OPClient into components/op.ts
components/op.ts and components/store/op.ts were near-verbatim copies of the
same 1Password Connect client, both live: the store copy served
components/store/{index,interfaces}.ts while the root copy served the Pulumi
dynamic provider and several components. Every future change had to be made
twice.

Keep components/op.ts as the single implementation:

- re-export TypeEnum / PurposeEnum / CategoryEnum, which previously only the
  store copy provided
- adopt the store copy's fuller comment on the byte-stable sort in
  findItemsByTag
- retarget components/store/{index,interfaces}.ts at ../op.ts
- delete components/store/op.ts

removeUndefinedProperties continues to come from components/helpers.ts rather
than the inlined duplicate the store copy carried.

No behavior change. This is groundwork for the OpenBao migration, where the
client is reimplemented against KV v2 — doing it once rather than twice.
